Become compatible with Python 2

+++ issues/tests/test_pagination.py
@@ -1,3 +1,4 @@
+from __future__ import division
 import pytest
 from math import ceil
 
@@ -15,6 +16,6 @@ def test_pagination(mf_api_client, random_service, param, count):
         Issue.objects.create(service=random_service, description=i)
     resp = mf_api_client.get(ISSUE_LIST_ENDPOINT, {param: count})
     assert resp['x-result-count'] == str(30)
-    assert resp['x-page-count'] == str(ceil(30 / count))
+    assert resp['x-page-count'] == str(int(ceil(30 / count)))
     content = get_data_from_response(resp, schema=LIST_OF_ISSUES_SCHEMA)
     assert len(content) == count
